Move FilterBuilderService into sub-package.
