+++ service/src/main/java/bio/terra/tanagra/service/export/DataExportService.java
@@ -8,12 +8,12 @@ import bio.terra.tanagra.app.configuration.ExportConfiguration;
 import bio.terra.tanagra.app.configuration.ExportConfiguration.PerModel;
 import bio.terra.tanagra.app.configuration.FeatureConfiguration;
 import bio.terra.tanagra.filterbuilder.EntityOutput;
-import bio.terra.tanagra.service.FilterBuilderService;
 import bio.terra.tanagra.service.UnderlayService;
 import bio.terra.tanagra.service.artifact.ActivityLogService;
 import bio.terra.tanagra.service.artifact.CohortService;
 import bio.terra.tanagra.service.artifact.ReviewService;
 import bio.terra.tanagra.service.artifact.model.Cohort;
+import bio.terra.tanagra.service.filter.FilterBuilderService;
 import bio.terra.tanagra.underlay.entitymodel.Attribute;
 import bio.terra.tanagra.utils.RandomNumberGenerator;
 import com.google.common.annotations.VisibleForTesting;
